feat(coc): relocate Dreams provider activity into the Dreams tab (AC-05)
Move the "Dreams provider activity" queue + history out of the AI Provider
page and into the admin Dreams tab, reused rather than duplicated.

- Extract shared provider brand visuals (PROVIDER_LABELS, ProviderAvatar,
  brand SVG icons) into shared/providerVisuals.tsx so both the AI Provider
  page and the activity section share one source of truth.
- Move ProviderActivitySection into its own features/dreams component and
  render it from DreamsView (new optional providerActivity/error/onRefresh
  props). Refresh control preserved.
- Remove the section, its three props, and the now-dead providerActivity
  imports from AIProviderPage; AdminPanel passes the existing dream activity
  state/refresh to DreamsView instead, and the auto-refresh effect now loads
  dream activity when the Dreams tab is the active route (quota still loads
  on the AI Provider sub-tab).
- Tests: new DreamsView.test.tsx (attribution, empty, error, Refresh-preserved,
  no-handler) and a regression test that the section is gone from the AI
  Provider page. Update dashboard-spa knowledge reference.
