+++ .github/skills/coc-knowledge/references/dashboard-spa.md
@@ -521,13 +521,18 @@ The legacy "Tools" popover has been migrated into the Admin page's left
 sidebar, but there is no longer a generic Tools group. The Admin sidebar is
 grouped by user task: Configure, Knowledge, Connections, Operations, and
 Developer / Internals. Embedded tool rows keep stable ids (`memory-toggle`,
-`skills-toggle`, `logs-toggle`, `stats-toggle`,
+`skills-toggle`, `dreams-admin-toggle`, `logs-toggle`, `stats-toggle`,
 `servers-toggle`) and `data-tab` still carries the matching dashboard route;
-Servers is shown only when `isServersEnabled()` is true.
+Servers is shown only when `isServersEnabled()` is true. The Knowledge group's
+**Dreams** row (`dreams-admin-toggle`, route `#dreams-admin`) renders
+`features/dreams/DreamsView.tsx` and is the admin home for global Dreams config
+and the relocated **Dreams provider activity** queue + history section
+(`features/dreams/ProviderActivitySection.tsx`); that section no longer lives on
+the AI Provider page. It is distinct from the per-workspace `DreamsPanel`.
 
 Clicking an embedded tool row dispatches `SET_ACTIVE_TAB` and updates
 `location.hash` to the corresponding top-level route (`#memory`, `#skills`,
-`#logs`, `#stats`, `#servers`). The Router maps every embedded tool
+`#dreams-admin`, `#logs`, `#stats`, `#servers`). The Router maps every embedded tool
 tab plus `'admin'` itself to a single `<AdminPanel />` render, so the admin shell
 (sidebar + breadcrumb + right pane) stays mounted across navigation.
 `AdminPanel` switches on `state.activeTab` — when it matches an embedded tool
