+++ packages/coc/src/server/spa/client/react/admin/AIProviderPage.tsx
@@ -26,7 +26,7 @@ import {
     getTightestFiniteQuotaType,
     getUnlimitedQuotaTypes,
 } from '../shared/quotaUtils';
-import { formatProviderActivityTimeout, type AgentProviderWorkActivity } from '../shared/providerActivity';
+import { PROVIDER_LABELS, ProviderAvatar } from '../shared/providerVisuals';
 
 const ProviderModelsSection = lazy(() => import('../features/models/ProviderModelsSection').then(m => ({ default: m.ProviderModelsSection })));
 const ProviderEffortTiersSection = lazy(() => import('../features/models/ProviderEffortTiersSection').then(m => ({ default: m.ProviderEffortTiersSection })));
@@ -71,14 +71,10 @@ export interface AIProviderPageProps {
     quotaLoading: boolean;
     quotaError: string | null;
     onRefreshQuota: (options?: { force?: boolean }) => void;
-    providerActivity?: AgentProviderWorkActivity[];
-    providerActivityError?: string | null;
-    onRefreshProviderActivity?: () => void;
 
     sources: Record<string, string | undefined>;
 }
 
-const PROVIDER_LABELS: Record<Provider, string> = { copilot: 'Copilot', codex: 'Codex', claude: 'Claude' };
 const DEFAULT_PROVIDER_LABELS: Record<DefaultProvider, string> = PROVIDER_LABELS;
 const PROVIDER_IDS: Provider[] = ['copilot', 'codex', 'claude'];
 export const DEFAULT_AUTO_PROVIDER_ROUTING_CONFIG: NormalizedAutoProviderRoutingConfig = {
@@ -159,45 +155,6 @@ function formatCheckStatus(check: { status: string; reason: string } | undefined
     if (!check) return 'Not checked';
     const label = check.status.replace(/_/g, ' ');
     return `${label.charAt(0).toUpperCase()}${label.slice(1)} — ${check.reason}`;
-}
-
-function CopilotIcon() {
-    return (
-        <svg viewBox="0 0 256 208" width="18" height="14" fill="currentColor" aria-hidden="true">
-            <path d="M205.28 31.36c14.096 14.88 20.016 35.2 22.512 63.68c6.626 0 12.805 1.47 16.976 7.152l7.792 10.56A17.55 17.55 0 0 1 256 123.2v28.688c-.008 3.704-1.843 7.315-4.832 9.504C215.885 187.222 172.35 208 128 208c-49.066 0-98.19-28.273-123.168-46.608c-2.989-2.189-4.825-5.8-4.832-9.504V123.2c0-3.776 1.2-7.424 3.424-10.464l7.792-10.544c4.173-5.657 10.38-7.152 16.992-7.152c2.496-28.48 8.4-48.8 22.512-63.68C77.331 3.165 112.567.06 127.552 0H128c14.72 0 50.4 2.88 77.28 31.36m-77.264 47.376c-3.04 0-6.544.176-10.272.544c-1.312 4.896-3.248 9.312-6.08 12.128c-11.2 11.2-24.704 12.928-31.936 12.928c-6.802 0-13.927-1.42-19.744-5.088c-5.502 1.808-10.786 4.415-11.136 10.912c-.586 12.28-.637 24.55-.688 36.824c-.026 6.16-.05 12.322-.144 18.488c.024 3.579 2.182 6.903 5.44 8.384C79.936 185.92 104.976 192 128.016 192c23.008 0 48.048-6.08 74.512-18.144c3.258-1.48 5.415-4.805 5.44-8.384c.317-18.418.062-36.912-.816-55.312h.016c-.342-6.534-5.648-9.098-11.168-10.912c-5.82 3.652-12.927 5.088-19.728 5.088c-7.232 0-20.72-1.728-31.936-12.928c-2.832-2.816-4.768-7.232-6.08-12.128a106 106 0 0 0-10.24-.544m-26.941 43.93c5.748 0 10.408 4.66 10.408 10.409v19.183c0 5.749-4.66 10.409-10.408 10.409s-10.408-4.66-10.408-10.409v-19.183c0-5.748 4.66-10.408 10.408-10.408m53.333 0c5.749 0 10.409 4.66 10.409 10.409v19.183c0 5.749-4.66 10.409-10.409 10.409c-5.748 0-10.408-4.66-10.408-10.409v-19.183c0-5.748 4.66-10.408 10.408-10.408M81.44 28.32c-11.2 1.12-20.64 4.8-25.44 9.92c-10.4 11.36-8.16 40.16-2.24 46.24c4.32 4.32 12.48 7.2 21.28 7.2c6.72 0 19.52-1.44 30.08-12.16c4.64-4.48 7.52-15.68 7.2-27.04c-.32-9.12-2.88-16.64-6.72-19.84c-4.16-3.68-13.6-5.28-24.16-4.32m68.96 4.32c-3.84 3.2-6.4 10.72-6.72 19.84c-.32 11.36 2.56 22.56 7.2 27.04c10.56 10.72 23.36 12.16 30.08 12.16c8.8 0 16.96-2.88 21.28-7.2c5.92-6.08 8.16-34.88-2.24-46.24c-4.8-5.12-14.24-8.8-25.44-9.92c-10.56-.96-20 .64-24.16 4.32" />
-        </svg>
-    );
-}
-
-function OpenAIIcon() {
-    return (
-        <svg viewBox="140 140 520 520" width="18" height="18" fill="currentColor" aria-hidden="true">
-            <path d="m617.24 354a126.36 126.36 0 0 0-10.86-103.79 127.8 127.8 0 0 0-137.65-61.32 126.36 126.36 0 0 0-95.31-42.49 127.81 127.81 0 0 0-121.92 88.49 126.4 126.4 0 0 0-84.5 61.3 127.82 127.82 0 0 0 15.72 149.86 126.36 126.36 0 0 0 10.86 103.79 127.81 127.81 0 0 0 137.65 61.32 126.36 126.36 0 0 0 95.31 42.49 127.81 127.81 0 0 0 121.96-88.54 126.4 126.4 0 0 0 84.5-61.3 127.82 127.82 0 0 0-15.76-149.81zm-190.66 266.49a94.79 94.79 0 0 1-60.85-22c.77-.42 2.12-1.16 3-1.7l101-58.34a16.42 16.42 0 0 0 8.3-14.37v-142.39l42.69 24.65a1.52 1.52 0 0 1 .83 1.17v117.92a95.18 95.18 0 0 1-94.97 95.06zm-204.24-87.23a94.74 94.74 0 0 1-11.34-63.7c.75.45 2.06 1.25 3 1.79l101 58.34a16.44 16.44 0 0 0 16.59 0l123.31-71.2v49.3a1.53 1.53 0 0 1-.61 1.31l-102.1 58.95a95.16 95.16 0 0 1-129.85-34.79zm-26.57-220.49a94.71 94.71 0 0 1 49.48-41.68c0 .87-.05 2.41-.05 3.48v116.68a16.41 16.41 0 0 0 8.29 14.36l123.31 71.19-42.69 24.65a1.53 1.53 0 0 1-1.44.13l-102.11-59a95.16 95.16 0 0 1-34.79-129.81zm350.74 81.62-123.31-71.2 42.69-24.64a1.53 1.53 0 0 1 1.44-.13l102.11 58.95a95.08 95.08 0 0 1-14.69 171.55c0-.88 0-2.42 0-3.49v-116.68a16.4 16.4 0 0 0-8.24-14.36zm42.49-63.95c-.75-.46-2.06-1.25-3-1.79l-101-58.34a16.46 16.46 0 0 0-16.59 0l-123.31 71.2v-49.3a1.53 1.53 0 0 1 .61-1.31l102.1-58.9a95.07 95.07 0 0 1 141.19 98.44zm-267.11 87.87-42.7-24.65a1.52 1.52 0 0 1-.83-1.17v-117.92a95.07 95.07 0 0 1 155.9-73c-.77.42-2.11 1.16-3 1.7l-101 58.34a16.41 16.41 0 0 0-8.3 14.36zm23.19-50 54.92-31.72 54.92 31.7v63.42l-54.92 31.7-54.92-31.7z" />
-        </svg>
-    );
-}
-
-function ClaudeIcon() {
-    return (
-        <svg viewBox="0 0 100 100" width="18" height="18" fill="currentColor" aria-hidden="true">
-            <path d="m19.6 66.5 19.7-11 .3-1-.3-.5h-1l-3.3-.2-11.2-.3L14 53l-9.5-.5-2.4-.5L0 49l.2-1.5 2-1.3 2.9.2 6.3.5 9.5.6 6.9.4L38 49.1h1.6l.2-.7-.5-.4-.4-.4L29 41l-10.6-7-5.6-4.1-3-2-1.5-2-.6-4.2 2.7-3 3.7.3.9.2 3.7 2.9 8 6.1L37 36l1.5 1.2.6-.4.1-.3-.7-1.1L33 25l-6-10.4-2.7-4.3-.7-2.6c-.3-1-.4-2-.4-3l3-4.2L28 0l4.2.6L33.8 2l2.6 6 4.1 9.3L47 29.9l2 3.8 1 3.4.3 1h.7v-.5l.5-7.2 1-8.7 1-11.2.3-3.2 1.6-3.8 3-2L61 2.6l2 2.9-.3 1.8-1.1 7.7L59 27.1l-1.5 8.2h.9l1-1.1 4.1-5.4 6.9-8.6 3-3.5L77 13l2.3-1.8h4.3l3.1 4.7-1.4 4.9-4.4 5.6-3.7 4.7-5.3 7.1-3.2 5.7.3.4h.7l12-2.6 6.4-1.1 7.6-1.3 3.5 1.6.4 1.6-1.4 3.4-8.2 2-9.6 2-14.3 3.3-.2.1.2.3 6.4.6 2.8.2h6.8l12.6 1 3.3 2 1.9 2.7-.3 2-5.1 2.6-6.8-1.6-16-3.8-5.4-1.3h-.8v.4l4.6 4.5 8.3 7.5L89 80.1l.5 2.4-1.3 2-1.4-.2-9.2-7-3.6-3-8-6.8h-.5v.7l1.8 2.7 9.8 14.7.5 4.5-.7 1.4-2.6 1-2.7-.6-5.8-8-6-9-4.7-8.2-.5.4-2.9 30.2-1.3 1.5-3 1.2-2.5-2-1.4-3 1.4-6.2 1.6-8 1.3-6.4 1.2-7.9.7-2.6v-.2H49L43 72l-9 12.3-7.2 7.6-1.7.7-3-1.5.3-2.8L24 86l10-12.8 6-7.9 4-4.6-.1-.5h-.3L17.2 77.4l-4.7.6-2-2 .2-3 1-1 8-5.5Z" />
-        </svg>
-    );
-}
-
-const PROVIDER_ICONS: Record<Provider, () => JSX.Element> = {
-    copilot: CopilotIcon,
-    codex: OpenAIIcon,
-    claude: ClaudeIcon,
-};
-
-function ProviderAvatar({ provider }: { provider: Provider }) {
-    const Icon = PROVIDER_ICONS[provider];
-    return (
-        <span className={`aip-avatar aip-avatar-${provider}`} aria-hidden="true">
-            <Icon />
-        </span>
-    );
 }
 
 function StatusBadge({ available }: { available: boolean }) {
@@ -390,74 +347,6 @@ function AdminPercentInput({ value, onChange, testId }: {
     );
 }
 
-function ProviderActivitySection({
-    activity,
-    error,
-    onRefresh,
-}: {
-    activity: AgentProviderWorkActivity[];
-    error?: string | null;
-    onRefresh?: () => void;
-}) {
-    return (
-        <section className="aip-auto-card" aria-labelledby="provider-dream-activity-title" data-testid="provider-dream-activity">
-            <div className="aip-auto-head">
-                <div>
-                    <div className="aip-provider-line">
-                        <span className="aip-provider-main" id="provider-dream-activity-title">Dreams provider activity</span>
-                        <span className="aip-provider-source">queue + history</span>
-                    </div>
-                    <div className="aip-provider-meta">
-                        Active and recent Dream jobs are attributed to the provider, model, and timeout selected for each run.
-                    </div>
-                </div>
-                {onRefresh && (
-                    <button
-                        type="button"
-                        className="ar-btn ar-btn-secondary ar-btn-sm"
-                        onClick={onRefresh}
-                        data-testid="provider-dream-activity-refresh"
-                    >
-                        Refresh
-                    </button>
-                )}
-            </div>
-            {error ? (
-                <div className="aip-error-banner" data-testid="provider-dream-activity-error">⚠ {error}</div>
-            ) : activity.length === 0 ? (
-                <div className="aip-auto-disabled" data-testid="provider-dream-activity-empty">
-                    No active or recent Dreams work.
-                </div>
-            ) : (
-                <div className="aip-auto-rules">
-                    {activity.map(item => {
-                        const providerLabel = PROVIDER_LABELS[item.provider] ?? item.provider;
-                        const trigger = item.trigger === 'idle' ? 'Idle' : item.trigger === 'manual' ? 'Manual' : 'Dreams';
-                        const status = item.status ? item.status.replace(/-/g, ' ') : 'unknown';
-                        return (
-                            <div className="aip-auto-rule" key={item.id} data-testid={`provider-dream-activity-${item.id}`}>
-                                <div className="aip-auto-rule-main">
-                                    <ProviderAvatar provider={item.provider} />
-                                    <div>
-                                        <div className="aip-provider-line">
-                                            <span className="aip-provider-main">{item.label}</span>
-                                            <span className="ar-badge ar-badge-accent">{providerLabel}</span>
-                                        </div>
-                                        <div className="aip-provider-meta">
-                                            {trigger} · {status} · {item.model ?? 'provider default'} · {formatProviderActivityTimeout(item.timeoutMs)}
-                                        </div>
-                                        {item.error && <div className="aip-install-error">✕ {item.error}</div>}
-                                    </div>
-                                </div>
-                            </div>
-                        );
-                    })}
-                </div>
-            )}
-        </section>
-    );
-}
-
 type AIProviderSubTab = 'routing' | 'models';
 const AI_PROVIDER_SUBTABS: { id: AIProviderSubTab; label: string; icon: string }[] = [
     { id: 'routing', label: 'Provider routing', icon: '◇' },
@@ -473,7 +362,6 @@ export function AIProviderPage(props: AIProviderPageProps) {
         providerAvailability, sdkInstallStatuses, sdkInstallErrors, onInstallSdk,
         dirty, saving, onSave, onCancel,
         quotaData, quotaLoading, quotaError, onRefreshQuota,
-        providerActivity = [], providerActivityError, onRefreshProviderActivity,
     } = props;
 
     const normalizedAutoRouting = normalizeAutoProviderRoutingConfig(autoRoutingConfig);
@@ -898,11 +786,6 @@ export function AIProviderPage(props: AIProviderPageProps) {
                         </tbody>
                     </table>
                 </div>
-                <ProviderActivitySection
-                    activity={providerActivity}
-                    error={providerActivityError}
-                    onRefresh={onRefreshProviderActivity}
-                />
                 {/* Unavailability warnings */}
                 {defaultProvider === 'codex' && providerAvailability['codex'] && !providerAvailability['codex'].available && (
                     <div className="aip-unavailable-banner" data-testid="codex-sdk-unavailable-banner">
